Trait xylem::BTreeMapSchemaExt

pub trait BTreeMapSchemaExt: Schema { }

Implement this trait for a schema type to support standard HashMap conversion.

This allows HashMap<K, V> to be converted from HashMap<K::From, V::From>, applying the conversion for K for k. The value argument is passed as-is for each value. No conversion arguments can be passed to the key type (the default value is always used).